requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
appdirs=1.4.4=pypi_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
cmake=3.26.3=pypi_0
docker-pycreds=0.4.0=pypi_0
filelock=3.12.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
idna=3.4=pypi_0
jinja2=3.1.2=pypi_0
ld_impl_linux-64=2.40=h41732ed_0
libffi=3.4.2=h7f98852_5
libgcc-ng=12.2.0=h65d4601_19
libgomp=12.2.0=h65d4601_19
libnsl=2.0.0=h7f98852_0
libsqlite=3.40.0=h753d276_1
libuuid=2.38.1=h0b41bf4_0
libzlib=1.2.13=h166bdaf_4
lit=16.0.2=pypi_0
markupsafe=2.1.2=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.3=h27087fc_1
networkx=3.1=pypi_0
numpy=1.24.3=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
openssl=3.1.0=hd590300_2
pathtools=0.1.2=pypi_0
pillow=9.5.0=pypi_0
pip=23.1.2=pyhd8ed1ab_0
protobuf=4.22.3=pypi_0
psutil=5.9.5=pypi_0
python=3.9.16=h2782a2a_0_cpython
pyyaml=6.0=pypi_0
readline=8.2=h8228510_1
requests=2.29.0=pypi_0
sentry-sdk=1.21.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=67.7.2=pyhd8ed1ab_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sympy=1.11.1=pypi_0
tk=8.6.12=h27826a3_0
torch=2.0.0=pypi_0
torchaudio=2.0.1=pypi_0
torchvision=0.15.1=pypi_0
tqdm=4.65.0=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2023c=h71feb2d_0
urllib3=1.26.15=pypi_0
wandb=0.15.0=pypi_0
wheel=0.40.0=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
